+++ Changelog.md
@@ -18,7 +18,8 @@
   pilot factional military ships.
 * The Info window's missions tab now displays the current objective
   according to the OSD.
-* Logo now lights up red for Autism Acceptance Month.
+* Logo now lights up red for Autism Acceptance Month and turns into a
+  rainbow for Queer Pride Month.
 
 ## 0.1.0 (unreleased)